Split larger classes from flow.scala into their own files
