+++ pyron-core/src/main/scala/com/cloudentity/pyron/domain/flow/FlowFailure.scala
@@ -1,0 +1,6 @@
+package com.cloudentity.pyron.domain.flow
+
+sealed trait FlowFailure
+case object RequestPluginFailure extends FlowFailure
+case object ResponsePluginFailure extends FlowFailure
+case object CallFailure extends FlowFailure
